pyproject.toml
[project]
name = "tweeter"
version = "0.1.0"
package-mode = false
description = ""
authors = [
    {name = "Devv-ictor",email = "dev.victor.alexandre@gmail.com"}
]
requires-python = ">=3.14,<3.15"
dependencies = [
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.1.0,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)"
]
